build fix.
svn:sync-xref-src-repo-rev: r76725

+++ src/VBox/VMM/VMMR3/PDM.cpp
@@ -1,4 +1,4 @@
-/* $Id: PDM.cpp 40405 2012-03-08 18:27:45Z knut.osmundsen@oracle.com $ */
+/* $Id: PDM.cpp 40406 2012-03-08 18:33:41Z knut.osmundsen@oracle.com $ */
 /** @file
  * PDM - Pluggable Device Manager.
  */
@@ -2509,24 +2509,24 @@ VMMR3_INT_DECL(bool) PDMR3TracingAreAll(PVM pVM, bool fEnabled)
 {
     for (PPDMDEVINS pDevIns = pVM->pdm.s.pDevInstances; pDevIns; pDevIns = pDevIns->Internal.s.pNextR3)
     {
-        if (pDevIns->fTraceing != fEnabled)
+        if (pDevIns->fTraceing != (uint32_t)fEnabled)
             return false;
 
         for (PPDMLUN pLun = pDevIns->Internal.s.pLunsR3; pLun; pLun = pLun->pNext)
             for (PPDMDRVINS pDrvIns = pLun->pTop; pDrvIns; pDrvIns = pDrvIns->Internal.s.pDown)
-                if (pDrvIns->fTraceing != fEnabled)
+                if (pDrvIns->fTraceing != (uint32_t)fEnabled)
                     return false;
     }
 
 #ifdef VBOX_WITH_USB
     for (PPDMUSBINS pUsbIns = pVM->pdm.s.pUsbInstances; pUsbIns; pUsbIns = pUsbIns->Internal.s.pNext)
     {
-        if (pUsbIns->fTraceing != fEnabled)
+        if (pUsbIns->fTraceing != (uint32_t)fEnabled)
             return false;
 
         for (PPDMLUN pLun = pUsbIns->Internal.s.pLuns; pLun; pLun = pLun->pNext)
             for (PPDMDRVINS pDrvIns = pLun->pTop; pDrvIns; pDrvIns = pDrvIns->Internal.s.pDown)
-                if (pDrvIns->fTraceing != fEnabled)
+                if (pDrvIns->fTraceing != (uint32_t)fEnabled)
                     return false;
     }
 #endif
